# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Pick me UP API",
    "description": "Get your hands clean using it ;)",
    "termsOfService": ""
  },
  "basePath": "/",
...
  "schemes": [
    "http", "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/passenger": {
       "post": {
          "tags": ["Passenger"],
          "description": "Get Passengers by Id",
          "operationId": "for the URL",
          "consumes": [
          "application/json"
        	],
          "produces": [
            "application/json"
          ],
          "parameters": [
              	{
	                "name": "body",
	                "in": "body",
	                "description": "Get passenger by id",
	              	"schema": 
	              		{
		                	"$ref": "#/definitions/createPassenger"
		                }
        		}
          ],
          "responses": {
            "201": {
              "description": "create response",
              "schema": {
                 "$ref": "#/definitions/201success"
              }
            }
          }
        }
    },
...
    "/api/routes": {
       "post": {
          "tags": ["Routes"],
          "description": "Get Passengers by Id",
          "operationId": "for the URL",
          "consumes": [
          "application/json"
        	],
          "produces": [
            "application/json"
          ],
          "parameters": [
              	{
	                "name": "body",
	                "in": "body",
	                "description": "Get passenger by id",
	              	"schema": 
	              		{
		                	"$ref": "#/definitions/createRoute"
		                }
        		}
          ],
          "responses": {
            "201": {
              "description": "create response",
              "schema": {
                 "$ref": "#/definitions/201success"
              }
            }
          }
        }
    },
...
  },
  "definitions": {
...
 	  "createPassenger": {
      "properties": 
  		{	
	        "username": {
	          "required":true,
	          "type": "string",
	          "example": "string"
	        },
	        "first_name": {
	          "required":true,
	          "type": "string",
	          "example": "string"
	        },
	        "last_name": {
	          "required":true,
	          "type": "string",
	          "example": "string"
	        },
	        "email": {
	          "required":true,
	          "type": "email",
	          "example": "mail@mail.com"
	        },
	        "skype_id": {
	          "required":true,
	          "type": "string",
	          "example": "string"
	        },
	        "cellphone": {
	          "required":true,
	          "type":"integer",
	          "format": "int64",
	          "example": 999999999
	        } 
      }
    },
    "createRoute": {
      "properties":
      {
        "departure": {
          "required":true,
          "type": "string",
          "example": "BELATRIX_LA_MOLINA"
        },
        "departureTime": {
          "required":true,
          "type": "string",
          "example": "1472765799931"
        },
        "destination": {
          "required":true,
          "type": "string",
          "example": "CHORRILLOS"
        },
        "cost": {
          "required":true,
          "type": "double",
          "example": "20.0"
        },
        "paymentType": {
          "required":true,
          "type": "string",
          "example": "CASH"
        },
        "routeOwner": {
          "required":true,
          "type":"integer",
          "format": "int64",
          "example": 1
        },
        "addressDestination": {
          "required":true,
          "type": "string",
          "example": "[Javier Prado, Arequipa, Colonial]"
        },
        "passengers": {
          "required":true,
          "type": "string",
          "example": [2, 3]
        }
      }
    },
...
    "201success" : {
       "properties": {
        "message": {
          "type": "status",
          "example" : "200"
        }
      }
    },
...
  }
}